completed player move broadcast feature
