+++ server/src/socket/index.js
@@ -1,7 +1,7 @@
 // Custom Modules
 const join_room_handler = require("./join_room_handler.socket.js");
 const leave_room_handler = require("./leave_room_handler.socket.js");
-const piece_moved_handler = require("./piece_moved_handler.js");
+const piece_moved_handler = require("./piece_moved_handler.socket.js");
 
 const socket = (socket) => {
   console.log("Connected succesfully to the socket ...");
