+++ server/src/socket/piece_moved_handler.socket.js
